+++ src/main/java/project/trendpick_pro/domain/product/entity/dto/response/ProductResponse.java
@@ -1,14 +1,13 @@
 package project.trendpick_pro.domain.product.entity.dto.response;
 
 import com.querydsl.core.annotations.QueryProjection;
-import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import project.trendpick_pro.domain.common.file.CommonFile;
 import project.trendpick_pro.domain.product.entity.Product;
-import project.trendpick_pro.domain.tag.entity.Tag;
+import project.trendpick_pro.domain.tags.tag.entity.Tag;
 
 import java.util.ArrayList;
 import java.util.List;
