+++ src/main/java/project/trendpick_pro/domain/tags/favoritetag/repository/FavoriteTagRepository.java
@@ -1,8 +1,8 @@
-package project.trendpick_pro.domain.favoritetag.repository;
+package project.trendpick_pro.domain.tags.favoritetag.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import project.trendpick_pro.domain.favoritetag.entity.FavoriteTag;
+import project.trendpick_pro.domain.tags.favoritetag.entity.FavoriteTag;
 import project.trendpick_pro.domain.member.entity.Member;
 
 import java.util.List;
