+++ src/main/java/project/trendpick_pro/domain/tags/tag/entity/dto/response/TagListResponse.java
@@ -1,7 +1,7 @@
-package project.trendpick_pro.domain.tag.entity.dto.response;
+package project.trendpick_pro.domain.tags.tag.entity.dto.response;
 
 import lombok.Getter;
-import project.trendpick_pro.domain.tag.entity.Tag;
+import project.trendpick_pro.domain.tags.tag.entity.Tag;
 
 import java.util.List;
